[Refactor] split logger.js from generate-app.js
