+++ package.json
@@ -4,6 +4,9 @@
   "description": "",
   "main": "index.js",
   "type": "module",
+  "exports": {
+    "./utils/*": "./utils/*.js"
+  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -16,7 +19,7 @@
     "react"
   ],
   "author": "",
-  "license": "ISC",
+  "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/modern-agile-team/dev-portfolio-app.git"
