+++ utils/index.js
@@ -1,0 +1,1 @@
+export { default as logger } from "./logger.js";
